+++ src/test/java/com/coniverse/dangjang/domain/intro/controller/IntroControllerTest.java
@@ -10,10 +10,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.test.web.servlet.ResultActions;
 
-import com.coniverse.dangjang.ControllerTest;
 import com.coniverse.dangjang.domain.intro.dto.IntroResponse;
 import com.coniverse.dangjang.domain.intro.service.IntroService;
 import com.coniverse.dangjang.global.exception.BadRequestException;
+import com.coniverse.dangjang.support.ControllerTest;
 
 /**
  * @author TEO
