+++ src/test/java/com/coniverse/dangjang/support/ControllerTest.java
@@ -1,4 +1,4 @@
-package com.coniverse.dangjang;
+package com.coniverse.dangjang.support;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
